+++ Frontend/src/components/SearchPageComponents/Filter/GeneralFilter/index.jsx
