+++ Frontend/src/components/SearchPageComponents/Filter/TeamsFilter/index.jsx
