+++ Frontend/src/components/SearchPageComponents/ResultStatus/index.jsx
