+++ Frontend/src/components/SearchPageComponents/SearchBar/index.jsx
