+++ Frontend/src/components/SearchPageComponents/SearchResultList/InstitutionCard/index.jsx
