+++ Frontend/src/components/SearchPageComponents/SearchResultList/LabeledLink/index.jsx
