+++ Frontend/src/components/SearchPageComponents/SearchResultList/ProjectCard/index.jsx
