+++ Frontend/src/components/SearchPageComponents/SearchResultList/SearchCard/index.jsx
