+++ Frontend/src/components/SearchPageComponents/SearchResultList/TeamCard/index.jsx
