+++ Frontend/src/components/SearchPageComponents/SearchResultList/index.jsx
