+++ Frontend/src/components/SearchPageComponents/SearchTabs/index.jsx
